// backend/src/modules/room/room.interface.ts
import { RoomStatus, RoomType } from "../../../generated/prisma/client";

// model Room {
//   id            String     @id @default(uuid()) @db.Uuid
//   flatId        String     @db.Uuid
//   name          String     @db.VarChar(100)
//   roomType      RoomType
//   monthlyRent   Decimal    @db.Decimal(10, 2)
//   dailyRent     Decimal?   @db.Decimal(10, 2)
//   maxOccupants  Int
//   availableBed  Int
//   furnished     Boolean    @default(false)
//   availableFrom DateTime
//   status        RoomStatus @default(AVAILABLE)
//   createdAt     DateTime   @default(now())
//   updatedAt     DateTime   @updatedAt



//   flat            Flat         @relation(fields: [flatId], references: [id], onDelete: Cascade)
//   bookings        Booking[]
//   monthlyPayments MonthlyPay[]

//   @@index([flatId])
//   @@index([status])
//   @@index([roomType])
//   @@index([availableFrom])
//   @@map("rooms")
// }


export interface IRoom {
  id?: string;
  flatId?: string;
  name: string;
  roomType: RoomType;
  monthlyRent: number;
  dailyRent?: number;
  maxOccupants: number;
  availableBed: number;
  furnished: boolean;
  availableFrom: Date;
  status?: RoomStatus;
}